modulelist bugfix

+++ train/model_deg.py
@@ -18,12 +18,12 @@ def calc_deg(dataset: str, workdir: str) -> torch.Tensor:
 # Compute the maximum in-degree in the training data.
     max_degree = -1
     for data in train_dataset:
-        d = degree(data.edge_index[1].to(torch.int64), num_nodes=data.num_nodes, dtype=torch.int32)
+        d = degree(data.edge_index[1], num_nodes=data.num_nodes, dtype=torch.long)
         max_degree = max(max_degree, int(d.max()))
 
 # Compute the in-degree histogram tensor
-    deg = torch.zeros(max_degree + 1, dtype=torch.int32)
+    deg = torch.zeros(max_degree + 1, dtype=torch.long)
     for data in train_dataset:
-        d = degree(data.edge_index[1].to(torch.int64), num_nodes=data.num_nodes, dtype=torch.int32)
+        d = degree(data.edge_index[1], num_nodes=data.num_nodes, dtype=torch.long)
         deg += torch.bincount(d, minlength=deg.numel())
     return deg
